requirements.txt
google-generativeai==0.8.3
pdf2image>=1.16.0
Pillow>=10.0.0
python-dotenv>=1.0.0
